refactor:将所有接口统一整理到cmdtype.d.ts中
