+++ app/components/common/dicer/cmd.ts
@@ -1,45 +1,3 @@
-import type { ExecutorProp } from "@/components/common/dicer/cmdPre";
-
-interface RoleAbility { // eslint-disable-line ts/consistent-type-definitions
-  abilityId?: number;
-  roleId?: number;
-  ruleId?: number;
-  act?: Record<string, string>;
-  ability?: Record<string, number>;
-}
-
-interface UserRole { // eslint-disable-line ts/consistent-type-definitions
-  userId: number;
-  roleId: number;
-  roleName?: string;
-  description?: string;
-  avatarId?: number;
-  state?: number;
-  modelName?: string;
-  speakerName?: string;
-  createTime?: string;
-  updateTime?: string;
-}
-
-interface CommandInfo { // eslint-disable-line ts/consistent-type-definitions
-  name: string;
-  alias: string[];
-  description: string;
-  examples: string[];
-  usage: string;
-}
-
-interface CPI { // eslint-disable-line ts/consistent-type-definitions
-  // 发送消息
-  sendMsg: (prop: ExecutorProp, msg: string) => void;
-  // 获取角色能力列表
-  getRoleAbilityList: (roleId: number) => Promise<RoleAbility>;
-  // 设置角色能力列表
-  setRoleAbilityList: (roleId: number, ability: RoleAbility) => Promise<void>;
-  // 发送Toast消息
-  sendToast: (msg: string) => void;
-}
-
 /**
  * 规则命名空间类，用于管理和执行一组相关命令
  *
