+++ app/components/common/dicer/cmdType.d.ts
@@ -1,11 +1,44 @@
-import type { ExecutorProp } from "@/components/common/dicer/cmdPre";
-
 interface CommandInfo { // eslint-disable-line ts/consistent-type-definitions
   name: string;
   alias: string[];
   description: string;
   examples: string[];
   usage: string;
+}
+
+/**
+ * 聊天室部分调用executor的时候, 会传这么一个结构体进去
+ */
+interface ExecutorProp { // eslint-disable-line ts/consistent-type-definitions
+  /**
+   * 房间ID
+   */
+  roomId?: number;
+  /**
+   * 完整的原始消息
+   */
+  originMessage?: string;
+  /**
+   * 指令消息id;
+   * 用于后续的指令回复,避免消息混乱;
+   */
+  replyMessageId?: number;
+  /**
+   * 骰娘的角色ID
+   */
+  dicerRoleId?: number;
+  /**
+   * 骰娘的头像ID
+   */
+  dicerAvatarId?: number;
+  /**
+   * 命令的主体, 不带前置的标点, 即中英文句号，也不包含@的人
+   */
+  command: string;
+  /**
+   * 聊天框中@的角色
+   */
+  mentionedRoles?: UserRole[];
 }
 
 interface CPI { // eslint-disable-line ts/consistent-type-definitions
@@ -17,6 +50,35 @@ interface CPI { // eslint-disable-line ts/consistent-type-definitions
   setRoleAbilityList: (roleId: number, ability: RoleAbility) => Promise<void>;
   // 发送Toast消息
   sendToast: (msg: string) => void;
+}
+
+interface RoleAbility { // eslint-disable-line ts/consistent-type-definitions
+  abilityId?: number;
+  roleId?: number;
+  ruleId?: number;
+  act?: Record<string, string>;
+  ability?: Record<string, number>;
+}
+
+interface UserRole { // eslint-disable-line ts/consistent-type-definitions
+  userId: number;
+  roleId: number;
+  roleName?: string;
+  description?: string;
+  avatarId?: number;
+  state?: number;
+  modelName?: string;
+  speakerName?: string;
+  createTime?: string;
+  updateTime?: string;
+}
+
+interface CommandInfo { // eslint-disable-line ts/consistent-type-definitions
+  name: string;
+  alias: string[];
+  description: string;
+  examples: string[];
+  usage: string;
 }
 
 /**
